// src/generated/models/Afb_dossierkypkysesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Afb_dossierkypkysesafb_statutdudossier = {
  0: 'Valid_',
  1: 'Enrevue',
  2: '_compl_ter',
  747010001: 'Suspendu',
  747010002: 'Rejet_'
} as const;
export type Afb_dossierkypkysesafb_statutdudossier = keyof typeof Afb_dossierkypkysesafb_statutdudossier;
export const Afb_dossierkypkysesstatecode = {
  0: 'Actif',
  1: 'Inactif'
} as const;
export type Afb_dossierkypkysesstatecode = keyof typeof Afb_dossierkypkysesstatecode;
export const Afb_dossierkypkysesstatuscode = {
  1: 'Actif',
  2: 'Inactif'
} as const;
export type Afb_dossierkypkysesstatuscode = keyof typeof Afb_dossierkypkysesstatuscode;

export interface Afb_dossierkypkysesBase {
  afb_commentairedconf?: string;
  afb_datededernierevalidation?: string;
  afb_datedesoumission?: string;
  afb_dossierkypkysid: string;
  "afb_nomdutiers@odata.bind"?: string;
  afb_prochainecheancier?: string;
  afb_referencedudossier: string;
  afb_statutdudossier: Afb_dossierkypkysesafb_statutdudossier;
  afb_tauxdecompletude: number;
  afb_versiondudossier: number;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Afb_dossierkypkysesstatecode;
  statuscode?: Afb_dossierkypkysesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Afb_dossierkypkyses extends Afb_dossierkypkysesBase {
  afb_nomdutiersname?: string;
  afb_statutdudossiername?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  afb_nomdutiers?: object;
  _afb_nomdutiers_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}