// src/generated/models/Afb_tiersesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Afb_tiersesafb_directionporteuse = {
  0: 'TRESO',
  1: 'DCONF',
  2: 'DMG',
  747010001: 'COMEX',
  747010002: 'DRISQUE'
} as const;
export type Afb_tiersesafb_directionporteuse = keyof typeof Afb_tiersesafb_directionporteuse;
export const Afb_tiersesafb_niveauderisque = {
  0: 'Critique',
  1: '_lev_',
  2: 'Standard'
} as const;
export type Afb_tiersesafb_niveauderisque = keyof typeof Afb_tiersesafb_niveauderisque;
export const Afb_tiersesafb_statutdutiers = {
  0: 'Partenaireactif',
  1: 'Cible',
  747010001: 'Suspendu',
  747010002: 'Rejet_',
  747010003: 'Cl_tur_'
} as const;
export type Afb_tiersesafb_statutdutiers = keyof typeof Afb_tiersesafb_statutdutiers;
export const Afb_tiersesstatecode = {
  0: 'Actif',
  1: 'Inactif'
} as const;
export type Afb_tiersesstatecode = keyof typeof Afb_tiersesstatecode;
export const Afb_tiersesstatuscode = {
  1: 'Actif',
  2: 'Inactif'
} as const;
export type Afb_tiersesstatuscode = keyof typeof Afb_tiersesstatuscode;

export interface Afb_tiersesBase {
  afb_adressecomplete?: string;
  "afb_chargederelation@odata.bind": string;
  afb_codeswiftbic?: string;
  afb_datedecreationsysteme: string;
  afb_datedernieremiseajour?: string;
  afb_directionporteuse: Afb_tiersesafb_directionporteuse;
  afb_emailcontactprincipal?: string;
  afb_niveauderisque: Afb_tiersesafb_niveauderisque;
  afb_nomdupartenaire: string;
  afb_numerorccmimmatriculation?: string;
  afb_pays: string;
  afb_statutdutiers: Afb_tiersesafb_statutdutiers;
  afb_telephone?: string;
  afb_tiersid: string;
  "afb_typejuridique@odata.bind": string;
  afb_ville?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Afb_tiersesstatecode;
  statuscode?: Afb_tiersesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Afb_tierses extends Afb_tiersesBase {
  afb_chargederelationname?: string;
  afb_directionporteusename?: string;
  afb_niveauderisquename?: string;
  afb_statutdutiersname?: string;
  afb_typejuridiquename?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  afb_chargederelation?: object;
  _afb_chargederelation_value?: string;
  afb_typejuridique?: object;
  _afb_typejuridique_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}